EctoSharding v0.0.1 EctoSharding

EctoSharding provides a shard registry service and an Ecto.Repo compatible query interface to support a sharded database structure.

Functions

current_shard(shard)

start_link()